Refactor and switch 'hub' type to v1alpha2
