+++ porch/controllers/packagevariantsets/api/v1alpha1/packagevariantset_conversion.go
@@ -14,4 +14,96 @@
 
 package v1alpha1
 
-func (*PackageVariantSet) Hub() {}
+import (
+	"fmt"
+
+	"sigs.k8s.io/controller-runtime/pkg/conversion"
+
+	pkgvarapi "github.com/GoogleContainerTools/kpt/porch/controllers/packagevariants/api/v1alpha1"
+	v1alpha2 "github.com/GoogleContainerTools/kpt/porch/controllers/packagevariantsets/api/v1alpha2"
+)
+
+func (src *PackageVariantSet) ConvertTo(dstRaw conversion.Hub) error {
+	dst := dstRaw.(*v1alpha2.PackageVariantSet)
+
+	return ConvertV1alpha1ToV1alpha2(src, dst)
+}
+
+func (dst *PackageVariantSet) ConvertFrom(srcRaw conversion.Hub) error {
+	src := srcRaw.(*v1alpha2.PackageVariantSet)
+
+	return ConvertV1alpha2ToV1alpha1(src, dst)
+}
+
+
+func ConvertV1alpha2ToV1alpha1(src *v1alpha2.PackageVariantSet, dst *PackageVariantSet) error {
+	dst.ObjectMeta = src.ObjectMeta
+
+	if src.Spec.Upstream != nil {
+		dst.Spec.Upstream = &Upstream{
+			Package: &Package{
+				Repo: src.Spec.Upstream.Repo,
+				Name: src.Spec.Upstream.Package,
+			},
+			Revision: src.Spec.Upstream.Revision,
+		}
+	}
+
+	// convert all the targets
+	for _, t := range src.Spec.Targets {
+		// Repositories map to Package targets
+		for _, r := range t.Repositories {
+			if len(r.PackageNames) == 0 {
+				pkg := src.ObjectMeta.Name
+				if src.Spec.Upstream != nil {
+					pkg = src.Spec.Upstream.Package
+				}
+				dst.Spec.Targets = append(dst.Spec.Targets, Target{
+					Package: &Package{
+						Repo: r.Name,
+						Name: pkg,
+					},
+				})
+			}
+			for _, pn := range r.PackageNames {
+				dst.Spec.Targets = append(dst.Spec.Targets, Target{
+					Package: &Package{
+						Repo: r.Name,
+						Name: pn,
+					},
+				})
+			}
+		}
+
+		// RepositorySelector maps directly
+		dst.Spec.Targets = append(dst.Spec.Targets, Target{Repositories: t.RepositorySelector})
+
+		// TODO: support this conversion
+		// just fail object selectors for now
+		if t.ObjectSelector != nil {
+			return fmt.Errorf("conversion of object selector targets is not supported")
+		}
+
+		// TODO: support this conversion
+		// just fail now (we should be able to convert some parts)
+		if t.Template != nil {
+			return fmt.Errorf("conversion of package variant template is not supported")
+		}
+	}
+
+	return nil
+}
+
+func ConvertV1alpha1ToV1alpha2(src *PackageVariantSet, dst *v1alpha2.PackageVariantSet) error {
+	dst.ObjectMeta = src.ObjectMeta
+
+	if src.Spec.Upstream != nil && src.Spec.Upstream.Package != nil {
+		dst.Spec.Upstream = &pkgvarapi.Upstream{
+			Repo:     src.Spec.Upstream.Package.Repo,
+			Package:  src.Spec.Upstream.Package.Name,
+			Revision: src.Spec.Upstream.Revision,
+		}
+	}
+
+	return nil
+}
